Tehre is noh time to beh lohst! (Blood Ravens.)
Batul Brothas!

Spehss Mahrens, todeh the enemeh is at oua doar! We know oua duteh and we will do eet. We fight for our honor as Blod Rehvens,
as SPESS MAHRENS, and we fight in the nehme of the Empra!

And if we die this deh we die in gloareh, we die heroes' deffs, but we shall not die, no! It is the enemeh who will tehste
deff and defeat!

As you know! Moast of oua battle brothars are shtehtioned in SPEHSS, Pruhpeared to deep strike! Oua perimeter has been
pruhpeared in the even dat oua enehmies should be so bald and so foolish. We have plehced numerous beacons, allowing for
muhltiple, simuln-tehneous and devashtehting defensive deep strikes

The Codecks astartees nehmes this maneuvah Steel Rehn. We will descend upon the foe, we will ovawhelm them - we will leave none
alive! Meanwhile oua ground fawses will ensue the full defense of oua headkwaters

We are the spehss mahrens! WE ARE THE EMPRA'S FUREH!