First wave of landmark icons
Some are better, some are not. I found no references to general and enforced pins so I went with something that seemed logical to me, a star and a enforcer respectively.

I also cried half of the time making these, mostly because I can't sprite a skull for shit so enjoy your Mayan God Cybernetica Acolyte.